Optimization no. 2 - use iovec
Since we setting TCP_NODELAY option on the socket, data is not
buffered anymore. This means that original approach that assumes
that we write the header and then move to the message payload
will result in 2 TCP messages instead desired one.

This commit uses writev to enable atomic non-contiguous data send.