mpi/pmix: Fix a typo in I/O engine code
Relase "bool" with "int" as the return type of `pmixp_io_fd()`. This
was causing an interesting hidden bug affecting the performance.
Since this function was returning boolean value it was always
returning "1" instead of actual fd number. fd=1 is set to /dev/null
for slurmstepd, this is a char device that is always read/write ready
from the poll() perspective. So poll was continuously interrupting and
progress was fine, but CPU usage was ~100%.